Improve MySQL routine and improve testing
For cases of dimensioning by week, month, and year (which would be
common from a practical sense), added code that would short-circuit
the logic and immediately return a value, as opposed to building a
result.

Improved testing of the implicit datetime groupings by adding data
spanning across the spectrum of testable time periods.